`[types]` Split out types to constituent files and modules.
