+++ frontend/src/old-pages/Clusters/Logs.tsx
@@ -8,6 +8,7 @@
 // or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 // OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
+import { LogEvent } from '../../types/logs'
 import React from 'react';
 import { useSearchParams } from "react-router-dom"
