+++ frontend/src/old-pages/Clusters/Properties.tsx
@@ -8,7 +8,7 @@
 // or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 // OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
-import { ClusterStatus } from '../../types/constants'
+import { ClusterStatus, ClusterDescription } from '../../types/clusters'
 import React from 'react';
 
 import { findFirst, clusterDefaultUser } from '../../util'
