+++ frontend/src/old-pages/Clusters/Scheduling.tsx
@@ -8,6 +8,7 @@
 // or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 // OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
+import { JobSummary, Job } from '../../types/jobs'
 import React from 'react';
 
 import { clearState, consoleDomain, getState, setState, ssmPolicy, useState } from '../../store'
