+++ frontend/src/old-pages/CustomImages/CustomImageDetails.tsx
@@ -8,7 +8,7 @@
 // or in the "LICENSE.txt" file accompanying this file. This file is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES
 // OR CONDITIONS OF ANY KIND, express or implied. See the License for the specific language governing permissions and
 // limitations under the License.
-import { ImageBuildStatus } from '../../types/constants'
+import { ImageBuildStatus } from '../../types/images'
 import React, { useCallback } from 'react';
 
 import { setState, getState, useState } from '../../store'
